fix: moved ResourceCard back to UI folder, because otherwise feature 79 won't work in this sprint, it has to be moved in the next sprint
